+++ parser/clj_kondo/impl/rewrite_clj/parser/token.clj
@@ -1,7 +1,10 @@
 (ns ^{:no-doc true} clj-kondo.impl.rewrite-clj.parser.token
   (:require [clj-kondo.impl.rewrite-clj
              [node :as node]
-             [reader :as r]]))
+             [reader :as r]]
+            [clj-kondo.impl.toolsreader.v1v2v2.clojure.tools.reader.reader-types :as rt]))
+
+(def ^:dynamic *invalid-token-exceptions* nil)
 
 (defn- read-to-boundary
   [reader & [allowed]]
@@ -36,12 +39,23 @@
 (defn parse-token
   "Parse a single token."
   [reader]
-  (let [first-char (r/next reader)
-        s (->> (if (= first-char \\)
-                 (read-to-char-boundary reader)
-                 (read-to-boundary reader))
-               (str first-char))
-        v (r/string->edn s)]
-    (if (symbol? v)
-      (symbol-node reader v s)
-      (node/token-node v s))))
+  (try
+    (let [first-char (r/next reader)
+          s (->> (if (= first-char \\)
+                   (read-to-char-boundary reader)
+                   (read-to-boundary reader))
+                 (str first-char))
+          v (r/string->edn s)]
+      (if (symbol? v)
+        (symbol-node reader v s)
+        (node/token-node v s)))
+    (catch Exception e
+      (let [{:keys [:type :ex-kind]} (ex-data e)]
+        (when (and *invalid-token-exceptions*
+                   (= :reader-exception type)
+                   (= :reader-error ex-kind))
+          (let [f {:row (rt/get-line-number reader)
+                   :col (rt/get-column-number reader)
+                   :message (.getMessage e)}]
+            (swap! *invalid-token-exceptions* conj (ex-info "Syntax error" {:findings [f]})))))
+      reader)))