+++ src/clj_kondo/impl/analyzer/match.clj
@@ -24,12 +24,49 @@
             (recur (next children) bindings)))
         bindings))))
 
+(defn analyze-list [ctx expr]
+  (let [children (:children expr)
+        snd (second children)]
+    (if-let [k (:k snd)]
+      (if (or (identical? :<< k)
+              (identical? :when k)
+              (identical? :guard k))
+        ;; https://github.com/clojure/core.match/blob/fb3188934ab9b6df0249ba3092a888def3434eee/src/main/clojure/clojure/core/match.clj#L1835
+        (let [bnds (analyze-expr ctx (first children))]
+          (common/analyze-expression** ctx (first (nnext children)))
+          bnds)
+        (analyze-children ctx expr))
+      (analyze-children ctx expr))))
+
+(defn analyze-vector [ctx expr]
+  (let [children (:children expr)]
+    (loop [children (seq children)
+           bindings {}]
+      (if children
+        (let [[child maybe-op & rchildren] children
+              k (:k maybe-op)]
+          (if (and k (or (identical? :when k)
+                         (identical? :guard k)))
+            ;; flattened syntax
+            (let [bnds (analyze-expr ctx child)]
+              (common/analyze-expression** ctx (first rchildren))
+              (merge bnds (analyze-vector ctx {:children (rest rchildren)})))
+            (if-let [bnds (analyze-expr ctx child)]
+              (recur (next children)
+                     (into bindings bnds))
+              (recur (next children) bindings))))
+        bindings))))
+
 (defn analyze-expr [ctx expr]
   (let [tag (utils/tag expr)]
     (case tag
       :token
       (analyze-token ctx expr)
-      (:list :vector :map)
+      (:list)
+      (analyze-list ctx expr)
+      (:vector)
+      (analyze-vector ctx expr)
+      (:map)
       (analyze-children ctx expr)
       ;; default
       (do (common/analyze-expression** ctx expr)
@@ -44,18 +81,3 @@
                   (utils/ctx-with-bindings ctx bindings)
                   ctx)]
         (common/analyze-expression** ctx ret)))))
-
-;; TODO:
-
-;; defpred
-;; (defpred even?)
-;; (defpred odd?)
-;; (defpred div3?)
-
-;; (deftest guard-pattern-match-1
-;;   (is (= (let [y '(2 3 4 5)]
-;;            (match [y]
-;;                   [([_ (a :when even?) _ _] :seq)] :a0
-;;                   [([_ (b :when [odd? div3?]) _ _] :seq)] :a1
-;;                   :else []))
-;;          :a1)))